test: add original-vs-port comparison framework
